move IData to chat folder
